+++ src/env.ts
@@ -17,8 +17,7 @@ const float = (n) => {
 }
 
 const VARS = {
-  adminInterfaceId: process.env['TRYORAMA_ADMIN_INTERFACE_ID'] || 'tryorama-admin-interface',
-  zomeInterfaceId: process.env['TRYORAMA_ZOME_INTERFACE_ID'] || 'tryorama-zome-interface',
+  interfaceId: process.env['TRYORAMA_INTERFACE_ID'] || 'tryorama-interface',
   stateDumpOnError: process.env['TRYORAMA_STATE_DUMP'] || true,
   zomeCallTimeoutMs: int(process.env['TRYORAMA_ZOME_CALL_TIMEOUT_MS']) || 90000,
   conductorTimeoutMs: int(process.env['TRYORAMA_CONDUCTOR_TIMEOUT_MS']) || 125000,